Update the general part of the comparison
Updated it because alternative.to still refers to https://android.biglybt.com/ which links to https://biglysoftware.github.io/BiglyBT-Android/TorrentAppComparison which is this file and is pretty outdated. I updated it with info from the play store site. Where it said size varies with device I used the size the play store page on my phone said.  I only updated the general part and hope someone else will be able to fix the rest. I also added libretorrent because it is a popular open source android torrent client and number 6 on the play store.